Fix ProcessExecutor bootstrapping in monorepo-helper
Without this fix Composer throws a LogicException with the message: "You
must use the ProcessExecutor instance which is part of a Composer\Loop
instance to be able to run async processes.", but only if the
monorepo-helper is not run on the `master` or `main` branch. The
inspiration for this fix was taken from the resolving commit of
composer/composer issue #10703.

See: https://github.com/composer/composer/issues/10703

Change-Id: I0af1c8f8f5badb0594c978a9c5dc8361e7bd6b81
Issue-Link: https://project.pronovix.net/issues/18528